Fixed: #128 User should be logged in to confirm email #126 Email confirm link with invalid username breaks app #127 Email confirm link with invalid key renders no error message
#128:
Instead of auto-logging in the user, we either redirect the user to the signin page if they're not logged in, or to their profile page. In this instance the login form includes a hidden field with the confirm_email key to pass once user logs in. Success or failure thus always shows successfully on their profile page.

#126:
Since we require login with #128, we can remove the username parameter from these requests entirely.

#127:
Since there's only one point of entry and no redirects now, users will always see a failed to confirm message or success message.